多项更新

+++ app/src/main/java/com/teamsolo/swear/structure/ui/konwledge/KnowledgeFragment.java
@@ -162,7 +162,7 @@ public class KnowledgeFragment extends HandlerFragment implements
 
         mSlideShow.post(() -> {
             int width = mSlideShow.getMeasuredWidth();
-            int aimHeight = width * 399 / 1080;
+            int aimHeight = (int) (width * 399.0 / 1080);
             ViewGroup.LayoutParams params = mSlideShow.getLayoutParams();
             if (params.height != aimHeight) {
                 params.height = aimHeight;
